src/evaluation/requirements.txt
google-cloud-bigquery
pandas
pandas-gbq
db-dtypes
pyarrow
